+++ lib/screen/SALES/sale_itemlist.dart
@@ -508,7 +508,8 @@ class _SalesItemState extends State<SalesItem> {
                                                                           ["tax"]),
                                                                   0.0,
                                                                   "0",
-                                                                  0,index);
+                                                                  0,
+                                                                  index);
 
                                                           if (result ==
                                                               "success") {
@@ -529,7 +530,7 @@ class _SalesItemState extends State<SalesItem> {
                                                                             index]
                                                                         [
                                                                         "code"],
-                                                                    int.parse(value
+                                                                    double.parse(value
                                                                         .qty[
                                                                             index]
                                                                         .text),
@@ -545,7 +546,8 @@ class _SalesItemState extends State<SalesItem> {
                                                                               index]
                                                                           [
                                                                           "tax"],
-                                                                    ),value.tax,
+                                                                    ),
+                                                                    value.tax,
                                                                     100,
                                                                     0.0,
                                                                     0,
@@ -828,17 +830,35 @@ class _SalesItemState extends State<SalesItem> {
                                                                           ["tax"]),
                                                                   0.0,
                                                                   "0",
-                                                                  0,index);
+                                                                  0,
+                                                                  index);
                                                           print(
                                                               "result----$result");
                                                           if (result ==
                                                               "success") {
+                                                            // value
+                                                            //         .discount_prercent[
+                                                            //             index]
+                                                            //         .text =
+                                                            //     value.disc_per
+                                                            //         .toString();
+                                                            // value
+                                                            //         .discount_amount[
+                                                            //             index]
+                                                            //         .text =
+                                                            //     value.disc_amt
+                                                            //         .toString();
+                                                            // value
+                                                            //         .salesqty[index]
+                                                            //         .text =
+                                                            //     value.qty[index]
+                                                            //         .text
+                                                            //         .toString();
                                                             // value.discount_prercent[index].text = value.disc_per.toString();
                                                             var res = await OrderAppDB
                                                                 .instance
                                                                 .insertsalesBagTable(
-                                                                    products[
-                                                                            index]
+                                                                    products[index]
                                                                         [
                                                                         "item"],
                                                                     s[0],
@@ -853,7 +873,7 @@ class _SalesItemState extends State<SalesItem> {
                                                                             index]
                                                                         [
                                                                         "code"],
-                                                                    int.parse(value
+                                                                    double.parse(value
                                                                         .qty[
                                                                             index]
                                                                         .text),
@@ -870,7 +890,6 @@ class _SalesItemState extends State<SalesItem> {
                                                                               index]
                                                                           [
                                                                           "tax"],
-                                                                          
                                                                     ),
                                                                     value.tax,
                                                                     100,
